requirements.txt
# Core LangChain packages
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0
langchain-google-genai>=0.0.5

# Vector databases and embeddings
chromadb>=0.4.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4

# Google Gemini integration
google-generativeai>=0.3.0
google-cloud-aiplatform>=1.38.0

# Authentication and environment management
python-dotenv>=1.0.0

# Document processing
pypdf>=3.17.0
python-docx>=0.8.11
python-pptx>=0.6.21
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Web scraping and data handling
requests>=2.31.0
aiohttp>=3.8.0
selenium>=4.15.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0

# Environment and configuration
python-dotenv>=1.0.0
pydantic-settings>=2.0.0

# Web framework (if building web app)
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.28.0

# Utilities
# tiktoken>=0.5.0
# tqdm>=4.66.0
# click>=8.1.0

# Testing and development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Optional: Advanced features
# redis>=5.0.0
# psycopg2-binary>=2.9.0
# sqlalchemy>=2.0.0